+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## xx.xx.xx
-* Views will now be stopped and started again when the app transitions between background and foreground, instead of being paused and resumed.
+## 24.7.2
+* Automatic view pause/resumes are changed with stop/start for better data consistency.
+* Added the config interface 'experimental' to group experimental features.
+* Added a flag (enableViewNameRecording) to add view names to each event. (Experimental!)
+* Added a flag (enableVisibilityTracking) to add app visibility info to views * Added Content feature methods:
+	- enterContentZone, to start Content checks(Experimental!)
+	- exitContentZone, to stop content checks (Experimental!)
 
 ## 24.7.1
 * Added `enableTemporaryDeviceIDMode` config and post-initialization methods to enable temporary device ID mode